+++ packages/desktop-client/src/components/common.tsx
@@ -28,6 +28,8 @@ export { default as Input } from './common/Input';
 export { default as InputWithContent } from './common/InputWithContent';
 export { default as Label } from './common/Label';
 export { default as Menu } from './common/Menu';
+export { default as MenuButton } from './common/MenuButton';
+export { default as MenuTooltip } from './common/MenuTooltip';
 export { default as Modal, ModalButtons } from './common/Modal';
 export { default as Search } from './common/Search';
 export { default as Select } from './common/Select';
