+++ upcoming-release-notes/1258.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [carkom]
+---
+
+Reorganized accounts directory. Pulled our Header functions to make the accounts.js smaller and more manageable.
